srsRAN: fix compilation for srsUE and and srsENB
The artifacts do not specify the compiler toolchain that is to be used
in order to reproduce the results. Apparently, gcc 13.2 is not the
used version, but we disabled the checks for it to finally compile.

In addition, the pdcp_entity_base was not even syntactically correct,
which makes reproduction even harder.

Signed-off-by: Eduard Vlad <eduard.vlad@rwth-aachen.de>